requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
python-dotenv>=1.0.0
grandalf>=0.8
flask>=3.0.0
requests>=2.31.0
